+++ src/api/operations.generated.ts
@@ -36,12 +36,9 @@ export const GQLOperations = {
     DisplayZone: 'DisplayZone',
     DisplayPlace: 'DisplayPlace',
     BudgetOverview: 'BudgetOverview',
-    Ceremony: 'Ceremony',
-    CeremonyCard: 'CeremonyCard',
     InternshipEngagementListItem: 'InternshipEngagementListItem',
     LanguageEngagementListItem: 'LanguageEngagementListItem',
     LanguageListItem: 'LanguageListItem',
-    MentorCard: 'MentorCard',
     MethodologiesCard: 'MethodologiesCard',
     OrganizationListItem: 'OrganizationListItem',
     PartnershipCard: 'PartnershipCard',
@@ -54,7 +51,10 @@ export const GQLOperations = {
     ProjectMemberList: 'ProjectMemberList',
     LoggedInUser: 'LoggedInUser',
     UserListItem: 'UserListItem',
+    Ceremony: 'Ceremony',
+    CeremonyCard: 'CeremonyCard',
     InternshipEngagementDetail: 'InternshipEngagementDetail',
+    MentorCard: 'MentorCard',
     LanguageEngagementDetail: 'LanguageEngagementDetail',
     LanguageDetail: 'LanguageDetail',
     LanguageForm: 'LanguageForm',
