+++ src/scenes/Engagement/CeremonyCard/CeremonyCard.generated.ts
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-duplicates, @typescript-eslint/no-empty-interface */
 import gql from 'graphql-tag';
-import * as Types from '../../api/schema.generated';
+import * as Types from '../../../api/schema.generated';
 
 export type CeremonyFragment = { __typename?: 'Ceremony' } & Pick<
   Types.Ceremony,
