+++ src/scenes/Engagement/CeremonyCard/CeremonyCard.stories.tsx
@@ -1,7 +1,7 @@
 import { action } from '@storybook/addon-actions';
 import { boolean, select } from '@storybook/addon-knobs';
 import React from 'react';
-import { dateTime } from '../knobs.stories';
+import { dateTime } from '../../../components/knobs.stories';
 import { CeremonyCard as Card } from './CeremonyCard';
 
 export default { title: 'Components' };
