+++ src/scenes/Engagement/InternshipEngagement/MentorCard/MentorCard.generated.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-duplicates, @typescript-eslint/no-empty-interface */
 import gql from 'graphql-tag';
-import * as Types from '../../api/schema.generated';
-import { UserListItemFragment } from '../UserListItemCard/UserListItem.generated';
-import { UserListItemFragmentDoc } from '../UserListItemCard/UserListItem.generated';
+import * as Types from '../../../../api/schema.generated';
+import { UserListItemFragment } from '../../../../components/UserListItemCard/UserListItem.generated';
+import { UserListItemFragmentDoc } from '../../../../components/UserListItemCard/UserListItem.generated';
 
 export type MentorCardFragment = { __typename?: 'SecuredUser' } & Pick<
   Types.SecuredUser,
