+++ src/scenes/Engagement/LanguageEngagement/LanguageEngagementDetail.generated.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-duplicates, @typescript-eslint/no-empty-interface */
 import gql from 'graphql-tag';
 import * as Types from '../../../api/schema.generated';
-import { CeremonyCardFragment } from '../../../components/CeremonyCard/CeremonyCard.generated';
-import { CeremonyCardFragmentDoc } from '../../../components/CeremonyCard/CeremonyCard.generated';
+import { CeremonyCardFragment } from '../CeremonyCard/CeremonyCard.generated';
+import { CeremonyCardFragmentDoc } from '../CeremonyCard/CeremonyCard.generated';
 
 export type LanguageEngagementDetailFragment = {
   __typename?: 'LanguageEngagement';
